+++ client/blocks/login/continue-as-user.jsx
@@ -43,10 +43,10 @@ function ContinueAsUser( {
 	redirectUrlFromQuery,
 	onChangeAccount,
 	redirectPath,
-	isSignUpFlow,
 	isWooOAuth2Client,
 	isWooPasswordless,
 	isBlazePro,
+	notYouText,
 } ) {
 	const translate = useTranslate();
 	const { url: validatedRedirectUrlFromQuery, loading: validatingQueryURL } =
@@ -63,26 +63,20 @@ function ContinueAsUser( {
 	// like that, but it is better than the alternative, and in practice it should happen quicker than
 	// the user can notice.
 
-	const translationComponents = {
-		br: <br />,
-		link: (
-			<button
-				type="button"
-				id="loginAsAnotherUser"
-				className="continue-as-user__change-user-link"
-				onClick={ onChangeAccount }
-			/>
-		),
-	};
-
-	const notYouText = isSignUpFlow
-		? translate( 'Not you?{{br/}} Sign out or log in with {{link}}another account{{/link}}', {
-				components: translationComponents,
-				args: { userName },
-				comment: 'Link to continue login as different user',
-		  } )
+	const notYouDisplayedText = notYouText
+		? notYouText
 		: translate( 'Not you?{{br/}}Log in with {{link}}another account{{/link}}', {
-				components: translationComponents,
+				components: {
+					br: <br />,
+					link: (
+						<button
+							type="button"
+							id="loginAsAnotherUser"
+							className="continue-as-user__change-user-link"
+							onClick={ onChangeAccount }
+						/>
+					),
+				},
 				args: { userName },
 				comment: 'Link to continue login as different user',
 		  } );
@@ -197,12 +191,12 @@ function ContinueAsUser( {
 				<Button
 					busy={ isLoading }
 					primary
-					href={ validatedRedirectUrlFromQuery || validatedRedirectPath || '/' }
+					href={ validatedRedirectPath || validatedRedirectUrlFromQuery || '/' }
 				>
 					{ translate( 'Continue' ) }
 				</Button>
 			</div>
-			<div className="continue-as-user__not-you">{ notYouText }</div>
+			<div className="continue-as-user__not-you">{ notYouDisplayedText }</div>
 		</div>
 	);
 }