+++ client/login/wp-login/index.jsx
@@ -25,6 +25,7 @@ import {
 	isA4AOAuth2Client,
 	isCrowdsignalOAuth2Client,
 	isWooOAuth2Client,
+	isGravatarOAuth2Client,
 } from 'calypso/lib/oauth2-clients';
 import { login, lostPassword } from 'calypso/lib/paths';
 import { addQueryArgs } from 'calypso/lib/url';
@@ -241,11 +242,14 @@ export class Login extends Component {
 	renderGravPoweredLoginBlockFooter() {
 		const { oauth2Client, translate, locale, currentQuery, currentRoute } = this.props;
 
+		const isGravatarMagicCode =
+			isGravatarOAuth2Client( oauth2Client ) && currentQuery.gravatar_magic_code === 'true';
 		const magicLoginUrl = login( {
 			locale,
 			twoFactorAuthType: 'link',
 			oauth2ClientId: currentQuery?.client_id,
 			redirectTo: currentQuery?.redirect_to,
+			gravatarMagicCode: isGravatarMagicCode,
 		} );
 		const currentUrl = new URL( window.location.href );
 		currentUrl.searchParams.append( 'lostpassword_flow', true );
@@ -268,7 +272,9 @@ export class Login extends Component {
 							this.props.recordTracksEvent( 'calypso_login_magic_login_request_click' )
 						}
 					>
-						{ translate( 'Email me a login link.' ) }
+						{ isGravatarMagicCode
+							? translate( 'Email me a login code.' )
+							: translate( 'Email me a login link.' ) }
 					</a>
 					<a
 						href={ lostPasswordUrl }