+++ client/my-sites/stats/features/modules/stats-referrers/stats-referrers.tsx
@@ -1,0 +1,98 @@
+import { StatsCard } from '@automattic/components';
+import { localizeUrl } from '@automattic/i18n-utils';
+import { megaphone } from '@wordpress/icons';
+import clsx from 'clsx';
+import { useTranslate } from 'i18n-calypso';
+import { useSelector } from 'react-redux';
+import QuerySiteStats from 'calypso/components/data/query-site-stats';
+import {
+	isRequestingSiteStatsForQuery,
+	getSiteStatsNormalizedData,
+} from 'calypso/state/stats/lists/selectors';
+import { getSelectedSiteId } from 'calypso/state/ui/selectors';
+import EmptyModuleCard from '../../../components/empty-module-card/empty-module-card';
+import { SUPPORT_URL } from '../../../const';
+import StatsModule from '../../../stats-module';
+import StatsModulePlaceholder from '../../../stats-module/placeholder';
+import { StatsEmptyActionSocial } from '../shared';
+
+type StatsRefeeresProps = {
+	className?: string;
+	period: string;
+	query: {
+		date: string;
+		period: string;
+	};
+	moduleStrings: {
+		title: string;
+		item: string;
+		value: string;
+		empty: string;
+	};
+};
+
+const StatsRefeeres: React.FC< StatsRefeeresProps > = ( {
+	period,
+	query,
+	moduleStrings,
+	className,
+} ) => {
+	const translate = useTranslate();
+	const siteId = useSelector( getSelectedSiteId ) as number;
+	const statType = 'statsReferrers';
+
+	// TODO: sort out the state shape.
+	const requesting = useSelector( ( state: any ) =>
+		isRequestingSiteStatsForQuery( state, siteId, statType, query )
+	);
+	const data = useSelector( ( state ) =>
+		getSiteStatsNormalizedData( state, siteId, statType, query )
+	) as [ id: number, label: string ]; // TODO: get post shape and share in an external type file.
+
+	return (
+		<>
+			{ siteId && statType && (
+				<QuerySiteStats statType={ statType } siteId={ siteId } query={ query } />
+			) }
+			{ /* This will be replaced with ghost loaders, fallback to the current implementation until then. */ }
+			{ requesting && <StatsModulePlaceholder isLoading={ requesting } /> }
+			{ ( ! data || ! data?.length ) && (
+				<StatsCard
+					className={ clsx( 'stats-card--empty-variant', className ) } // when removing stats/empty-module-traffic add this to the root of the card
+					title={ moduleStrings.title }
+					isEmpty
+					emptyMessage={
+						<EmptyModuleCard
+							icon={ megaphone }
+							description={ translate(
+								"We'll show you which websites are {{link}}referring visitors{{/link}} to your site.",
+								{
+									comment: '{{link}} links to support documentation.',
+									components: {
+										link: <a href={ localizeUrl( `${ SUPPORT_URL }#referrers` ) } />,
+									},
+									context: 'Stats: Info box label when the Referrers module is empty',
+								}
+							) }
+							cards={ <StatsEmptyActionSocial from="module_referrers" /> }
+						/>
+					}
+				/>
+			) }
+			{ /* TODO: consider supressing <StatsModule /> empty state */ }
+			{ data && !! data.length && (
+				<StatsModule
+					path="referrers"
+					moduleStrings={ moduleStrings }
+					period={ period }
+					query={ query }
+					statType={ statType }
+					showSummaryLink
+					className={ className } // TODO: extend with a base class after adding skeleton loaders
+				/>
+			) }
+		</>
+	);
+};
+
+export default StatsRefeeres;