+++ client/signup/steps/subscribe-email/content.jsx
@@ -1,3 +1,4 @@
+import { recordTracksEvent } from '@automattic/calypso-analytics';
 import { localize } from 'i18n-calypso';
 import SignupForm from 'calypso/blocks/signup-form';
 import ReskinnedProcessingScreen from 'calypso/signup/reskinned-processing-screen';
@@ -10,6 +11,8 @@ function SubscribeEmailStepContent( props ) {
 		handleCreateAccountError,
 		handleCreateAccountSuccess,
 		isPending,
+		redirectToAfterLoginUrl,
+		redirectToLogout,
 		redirectUrl,
 		step,
 		stepName,
@@ -31,17 +34,38 @@ function SubscribeEmailStepContent( props ) {
 				handleCreateAccountError={ handleCreateAccountError }
 				handleCreateAccountSuccess={ handleCreateAccountSuccess }
 				disableBlurValidation
-				disableContinueAsUser
 				isPasswordless
 				isReskinned
 				isSocialFirst={ false }
 				isSocialSignupEnabled={ false }
 				labelText={ translate( 'Your email' ) }
+				notYouText={ translate(
+					'Not you?{{br/}}Log out and {{link}}subscribe with %(email)s{{/link}}',
+					{
+						components: {
+							br: <br />,
+							link: (
+								<button
+									type="button"
+									id="subscribeDifferentEmail"
+									className="continue-as-user__change-user-link"
+									onClick={ () => {
+										recordTracksEvent( 'calypso_signup_click_on_change_account' );
+										redirectToLogout( window.location.href );
+									} }
+								/>
+							),
+						},
+						args: { email },
+						comment: 'Link to continue subscribe to email list as different user',
+					}
+				) }
 				queryArgs={ { user_email: email, redirect_to: redirectUrl } }
+				redirectToAfterLoginUrl={ redirectToAfterLoginUrl }
 				shouldDisplayUserExistsError
 				step={ step }
 				stepName={ stepName }
-				submitButtonLabel={ translate( 'Subscribe email' ) }
+				submitButtonLabel={ translate( 'Subscribe' ) }
 				submitButtonLoadingLabel={ translate( 'Subscribing…' ) }
 				suggestedUsername=""
 			/>