+++ packages/components/src/horizontal-bar-list/types.ts
@@ -59,7 +59,10 @@ type StatsActions = {
 };
 
 export type StatsCardProps = {
-	children: React.ReactNode;
+	/**
+	 * @property {React.ReactNode} children - the content of the card - not renders when isEmpty is true
+	 */
+	children?: React.ReactNode;
 	className?: string;
 	headerClassName?: string;
 	title: string;
@@ -69,8 +72,14 @@ export type StatsCardProps = {
 		label?: string;
 		url?: string;
 	};
+	/**
+	 * @property {boolean} isEmpty - renders an empty card with a message (`emptyMessage`) when true. It doesn't render column labels.
+	 */
 	isEmpty?: boolean;
 	isNew?: boolean;
+	/**
+	 * @property {string | React.ReactNode} emptyMessage - a message (or compoents) to display when the card is empty
+	 */
 	emptyMessage?: string | React.ReactNode;
 	/**
 	 * @property {string} metricLabel - a label to use for the values on the right side of the bars - `Views` by default