+++ packages/social-previews/src/threads-preview/post-preview.tsx
@@ -2,6 +2,7 @@ import { preparePreviewText } from '../helpers';
 import { Card } from './card';
 import { Footer } from './footer';
 import { Header } from './header';
+import { CAPTION_MAX_CHARS } from './helpers';
 import { Media } from './media';
 import { Sidebar } from './sidebar';
 import { ThreadsPreviewProps } from './types';
@@ -9,28 +10,19 @@ import { ThreadsPreviewProps } from './types';
 import './style.scss';
 
 export const ThreadsPostPreview: React.FC< ThreadsPreviewProps > = ( {
+	caption,
 	date,
-	description,
 	image,
 	media,
 	name,
 	profileImage,
 	showThreadConnector,
-	text,
 	title,
 	url,
 } ) => {
 	const hasMedia = !! media?.length;
 
 	const displayAsCard = url && image && ! hasMedia;
-
-	let textToDisplay = text || title;
-
-	// Attach the URL to the text if not displaying as a card and it's not already in the text.
-	textToDisplay =
-		! displayAsCard && textToDisplay && url && ! textToDisplay.includes( url )
-			? `${ textToDisplay } ${ url }`
-			: textToDisplay;
 
 	return (
 		<div className="threads-preview__wrapper">
@@ -39,20 +31,16 @@ export const ThreadsPostPreview: React.FC< ThreadsPreviewProps > = ( {
 				<div className="threads-preview__main">
 					<Header name={ name } date={ date } />
 					<div className="threads-preview__content">
-						{ textToDisplay ? (
+						{ caption ? (
 							<div className="threads-preview__text">
-								{ preparePreviewText( textToDisplay, { platform: 'threads' } ) }
+								{ preparePreviewText( caption, {
+									platform: 'threads',
+									maxChars: CAPTION_MAX_CHARS,
+								} ) }
 							</div>
 						) : null }
 						{ hasMedia ? <Media media={ media } /> : null }
-						{ displayAsCard ? (
-							<Card
-								description={ description || '' }
-								image={ image }
-								title={ title || '' }
-								url={ url }
-							/>
-						) : null }
+						{ displayAsCard ? <Card image={ image } title={ title || '' } url={ url } /> : null }
 					</div>
 					<Footer />
 				</div>